exploits/Inlook/requirements.txt
gmpy2==2.1.2
pycryptodome==3.16.0